[client] Allow creations of multiple FBOs and shaders
The GUI needs to be drawn to a separate FBO that won't be shaded using the primary shader.
Need to provide a way for the hooks to switch between active FBOs.

Will need to make a second pass-through shader eventually.
This change doesn't modify the names under which variables are exported to shaders.